chore(theme): renamed theme components to follow pascal case
